+++ udata/__init__.py
@@ -4,5 +4,5 @@
 udata
 '''
 
-__version__ = '6.0.1'
+__version__ = '6.0.2.dev'
 __description__ = 'Open data portal'