+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix date timezone and format for harvest previz [#2799](https://github.com/opendatateam/udata/pull/2799)
 - Add support for DCAT startDate and endDate in temporal coverage [#2801](https://github.com/opendatateam/udata/pull/2801)
 - New feature: Users can now change their email by themselves [#2792](https://github.com/opendatateam/udata/pull/2792)
+- Remove useless published date in resource model [#2807](https://github.com/opendatateam/udata/pull/2807)
 
 ## 6.0.0 (2023-01-09)